Refactorates messages folder
